requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-command
pulumi_openstack
python-openstackclient
python-designateclient
python-novaclient
